Slight refactor in component org
